+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.21.2
+
+* fixed that CPManager, this time with unit tests to back them up
+* added option to invalidate all connections on this CPManager